fix(ws3): use u64 trace totals and add metadata handler tests

+++ src/models/orchestration.rs
@@ -154,7 +154,7 @@ pub struct TraceResponse {
     pub run_id: String,
     pub events: Vec<TraceEvent>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub total: Option<usize>,
+    pub total: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub truncated: Option<bool>,
 }
